backend/python-mrz-service/requirements.txt
flask>=2.3.0
pillow>=9.0.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
